+++ x/foundation/keeper/member.go
@@ -16,7 +16,9 @@ func validateMetadata(metadata string, config foundation.Config) error {
 
 func (k Keeper) UpdateDecisionPolicy(ctx sdk.Context, policy foundation.DecisionPolicy) error {
 	info := k.GetFoundationInfo(ctx)
-	info.SetDecisionPolicy(policy)
+	if err := info.SetDecisionPolicy(policy); err != nil {
+		return err
+	}
 	info.Version++
 	k.SetFoundationInfo(ctx, info)